Fix vector angle calculation bug
Before, the angle_from() function was using the vector [0, 1]
to calculate the angle between a vector v and the x-axis. This
was supposed to be [1, 0]. This is the reason why passing vector
u2 was yielding the correct result. I fixed it to pass u1.
Moreover, instead of defaulting to [1, 0], I defaulted to
a vector e1, the first standard basis vector.

I had missed the part of the instruction that said to apply the
program to the matrices of exercise 1, so I added a loop to plot
each of these. Further, I added a command-line option to allow
the user to enter an image path, causing the progrma to additionally
save an image for each of the plot figures with that name pattern.